Fluid. It doesn't work well for me.
It causes the font sizing to go wonky. I could javascript the font size
larger as the screen grows, but I think that it'll be gratuitous and
distracting. So, moving back to a static layout should work great.

Need to add a media query to pop the font size up to 110% on widescreen
though. Otherwise it's golden.